+++ shared/packages/expectationManager/src/lib/lib.ts
@@ -1,0 +1,7 @@
+export function mapToObject<T>(map: Map<any, T>): { [key: string]: T } {
+	const o: { [key: string]: any } = {}
+	map.forEach((value, key) => {
+		o[key] = value
+	})
+	return o
+}