+++ tests/internal-tests/src/__mocks__/fs.ts
@@ -541,6 +541,16 @@ export function rename(source: string, destination: string, callback: (error: an
 	})
 }
 fs.rename = rename
+export function utimes(
+	_fileName: string,
+	_atime: Date,
+	_mtime: Date,
+	callback: (error: any, result?: any) => void
+): void {
+	// This is currently a no-op, since there isn't any support for atime/mtime in this mock
+	callback(undefined, undefined)
+}
+fs.rename = utimes
 
 export function createReadStream(path: string, _options?: BufferEncoding | undefined): FSReadStream {
 	return new FSReadStream(path)
@@ -640,6 +650,7 @@ fs.promises = {
 	open: promisify(open),
 	copyFile: promisify(copyFile),
 	rename: promisify(rename),
+	utimes: promisify(utimes),
 }
 
 module.exports = fs