+++ shared/packages/expectationManager/CHANGELOG.md
@@ -3,6 +3,55 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.50.2](https://github.com/nrkno/tv-automation-package-manager/compare/v1.50.2-alpha.3...v1.50.2) (2024-03-27)
+
+
+### Bug Fixes
+
+* (scaling); PM should not always scale up for all NEW or WAITING, only READY ([653adde](https://github.com/nrkno/tv-automation-package-manager/commit/653addebb087068ef568a0db46228ed11c4fdb5b))
+* ensure noWorkerAssignedTime is reset when there is an assigned worker ([c8266d6](https://github.com/nrkno/tv-automation-package-manager/commit/c8266d63939d21d57cde6f08f6e37e6435543c3e))
+* logic change: don't return from WAITING to NEW when the expectation isn't ready to be worked on. ([0b1036c](https://github.com/nrkno/tv-automation-package-manager/commit/0b1036cf920b11a9b553abec63d015e6c2d59b99))
+* update lastEvaluationTime in FULFILLED, so that we wait a bit longer before checking again. ([4a72b61](https://github.com/nrkno/tv-automation-package-manager/commit/4a72b6177c573bf6b7731e4e01708988d86a60c9))
+
+
+
+
+
+## [1.50.2-alpha.3](https://github.com/nrkno/tv-automation-package-manager/compare/v1.50.2-alpha.2...v1.50.2-alpha.3) (2024-03-25)
+
+**Note:** Version bump only for package @sofie-package-manager/expectation-manager
+
+
+
+
+
+## [1.50.2-alpha.2](https://github.com/nrkno/tv-automation-package-manager/compare/v1.50.2-alpha.1...v1.50.2-alpha.2) (2024-03-25)
+
+**Note:** Version bump only for package @sofie-package-manager/expectation-manager
+
+
+
+
+
+## [1.50.2-alpha.1](https://github.com/nrkno/tv-automation-package-manager/compare/v1.50.2-alpha.0...v1.50.2-alpha.1) (2024-03-25)
+
+**Note:** Version bump only for package @sofie-package-manager/expectation-manager
+
+
+
+
+
+## [1.50.2-alpha.0](https://github.com/nrkno/sofie-package-manager/compare/v1.50.1...v1.50.2-alpha.0) (2024-02-29)
+
+
+### Bug Fixes
+
+* fix an issue where workers wheren't scaled up properly when expectations where waiting ([81f354c](https://github.com/nrkno/sofie-package-manager/commit/81f354c301e587eba6f80b4f33b7fe3ebe14bf49))
+
+
+
+
+
 # [1.50.0](https://github.com/nrkno/tv-automation-package-manager/compare/v1.50.0-alpha.10...v1.50.0) (2024-02-19)
 
 **Note:** Version bump only for package @sofie-package-manager/expectation-manager